MAL: Fix the XML returned from the API
MAL appears to be HTML encoding the synopsis part of the XML returned
from the search API. This results in entirely invalid XML. The solution?
Re-encode it all to be valid UTF-8 XML. This approach uses the list of
HTML entities on Wikipedia to re-encode to UTF-8 where their list had
it, and the unicode escape sequence otherwise (e.g. &#160;)

tl;dr MAL sucks.